requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Click==7.0
defusedxml==0.6.0
Django==3.0.3
django-allauth==0.41.0
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-modeltranslation==0.14.4
django-recaptcha3==0.1.0
h11==0.9.0
httptools==0.5.0
idna==2.9
oauthlib==3.1.0
Pillow==9.5.0
python3-openid==3.1.0
pytz==2019.3
requests==2.28.2
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.0